+++ shardingsphere-governance/shardingsphere-governance-core/shardingsphere-governance-core-common/src/main/java/org/apache/shardingsphere/governance/core/yaml/config/schema/YamlTableMetaData.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.yaml.config.metadata;
+package org.apache.shardingsphere.governance.core.yaml.config.schema;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -24,11 +24,13 @@ import org.apache.shardingsphere.infra.yaml.config.YamlConfiguration;
 import java.util.Map;
 
 /**
- * ShardingSphere schema for YAML.
+ * Table meta data for YAML.
  */
 @Getter
 @Setter
-public final class YamlSchema implements YamlConfiguration {
+public final class YamlTableMetaData implements YamlConfiguration {
     
-    private Map<String, YamlTableMetaData> tables;
+    private Map<String, YamlColumnMetaData> columns;
+    
+    private Map<String, YamlIndexMetaData> indexes;
 }
